Set up tailwind, react-router-dom, refactor code so API info is accessible seperately through navbar
