+++ control-panel/index.html
@@ -4,7 +4,9 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite + React</title>
+    <title>Legion Control Panel</title>
+    <link href="/src/style.css" rel="stylesheet">
+
   </head>
   <body>
     <div id="root"></div>
