+++ control-panel/src/main.jsx
@@ -1,10 +1,22 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import React from 'react';
 import './index.css'
-import App from './App.jsx'
 
-createRoot(document.getElementById('root')).render(
+import  App  from './App'
+import API_Status from './pages/api_status.jsx';
+import Navbar from './components/Navbar.jsx';
+
+import ReactDOM from "react-dom/client";
+
+const root = document.getElementById("root");
+
+ReactDOM.createRoot(root).render(
   <StrictMode>
-    <App />
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
+
   </StrictMode>,
 )
