+++ control-panel/src/pages/home.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+const Home = () => {
+  return (
+        <h1 class = "text-3xl font-bold underline"> Hi </h1>
+
+  );
+}
+export default Home;
