+++ control-panel/src/pages/statistics.jsx
@@ -1,0 +1,5 @@
+export default function Statistics () {
+    return (
+        <h1 className="text-3xl font-bold underline"> Statistics Page </h1>
+    )
+}
