+++ control-panel/vite.config.js
@@ -1,11 +1,22 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react";
+import tailwindcss from "@tailwindcss/vite";
 
 export default defineConfig({
-  plugins: [react()],
+  plugins: [react(),
+    tailwindcss(),
+  ],
+  
   server: {
     proxy: {
-      '/api': 'http://127.0.0.1:5000'
-    }
-  }
-})
+      "/api": "http://127.0.0.1:5000",
+    },
+  },
+});
+
+
+
+
+
+
+
